% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sada_ee.R
\name{vcov.sada_ee}
\alias{vcov.sada_ee}
\title{Variance covariance matrix for SADA estimating equation fits}
\usage{
\method{vcov}{sada_ee}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"sada_ee"} as returned by
\code{sada_ee()}.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
A numeric matrix giving the estimated covariance of the SADA
estimator. This is the same matrix that is stored in the \code{vcov}
element of the object.
}
\description{
Extracts the estimated covariance matrix of the SADA estimator from an
object returned by \code{sada_ee}.
}
