% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sada_ols.R
\name{vcov.sada_ols}
\alias{vcov.sada_ols}
\title{Variance-covariance matrix for SADA OLS}
\usage{
\method{vcov}{sada_ols}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"sada_ols"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The covariance matrix of the coefficient estimates.
}
\description{
Variance-covariance matrix for SADA OLS
}
